shiny_app/requirements.txt
fastapi
pydantic
uvicorn
transformers
python-dotenv
streamlit
requests
langchain
langchain_groq
rsconnect-python
shiny